+++ Sources/Utilities/KMMultipleSelectionConfiguration.swift
@@ -17,6 +17,8 @@ public struct KMMultipleSelectionConfiguration {
     
     public var image : UIImage? =  UIImage(named: "checked", in: Bundle.km, compatibleWith: nil)
     
+    public var postSubmitImage : UIImage? = UIImage(named: "selectedCheck", in: Bundle.km, compatibleWith: nil)
+    
     public var backgroundColor: UIColor = UIColor.clear
     
     public var selectedBackgroundColor: UIColor = UIColor(hexString: "D4F3F8")
@@ -29,7 +31,11 @@ public struct KMMultipleSelectionConfiguration {
     
     public var borderColor: UIColor = UIColor(hexString: "00A4BF")
     
+    public var postSubmitBorderColor: UIColor = UIColor(hexString: "8F8F8F")
+    
     public var titleColor: UIColor = UIColor(hexString: "00A4BF")
+    
+    public var postSubmitTitleColor: UIColor = UIColor(hexString: "8F8F8F")
     
     public var topPadding : CGFloat = 12