+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
 ## Unreleased
+- Default configuration added for disabling the form submit button using 'disableFormPostSubmit'.
 - Added support of prefill checkboxes on Form Template.
 
 ## [1.2.1] 2023-11-02